Added file check for books/drugs/enginetweaks ini files
Fixed possible array index out of bound error in BarBoxes.
Added trimming to custom unarmed attack names.
Edits to ddraw.ini and function docs.

+++ sfall/Modules/ScriptExtender.cpp
@@ -933,7 +933,7 @@ void ScriptExtender::init() {
 	Worldmap::OnWorldmapLoop() += RunGlobalScriptsOnWorldMap;
 
 	globalScriptPathList = IniReader::GetConfigList("Scripts", "GlobalScriptPaths", "scripts\\gl*.int");
-	for (unsigned int i = 0; i < globalScriptPathList.size(); i++) {
+	for (size_t i = 0; i < globalScriptPathList.size(); i++) {
 		ToLowerCase(globalScriptPathList[i]);
 	}
